fixed some conversion warnings (LLVM)

+++ src/bmsse_util.h
@@ -75,7 +75,7 @@ void sse2_xor_arr_2_mask(__m128i* BMRESTRICT dst,
                          const __m128i* BMRESTRICT src_end,
                          bm::word_t mask)
 {
-     __m128i xM = _mm_set1_epi32(mask);
+     __m128i xM = _mm_set1_epi32((int)mask);
      do
      {
         _mm_store_si128(dst+0, _mm_xor_si128(_mm_load_si128(src+0), xM));
@@ -99,7 +99,7 @@ void sse2_andnot_arr_2_mask(__m128i* BMRESTRICT dst,
                             const __m128i* BMRESTRICT src_end,
                             bm::word_t mask)
 {
-     __m128i xM = _mm_set1_epi32(mask);
+     __m128i xM = _mm_set1_epi32((int)mask);
      do
      {
         _mm_store_si128(dst+0, _mm_andnot_si128(_mm_load_si128(src+0), xM)); // xmm1 = (~xmm1) & xM
@@ -474,7 +474,7 @@ void sse2_set_block(__m128i* BMRESTRICT dst,
                     __m128i* BMRESTRICT dst_end, 
                     bm::word_t value)
 {
-    __m128i xmm0 = _mm_set1_epi32(value);
+    __m128i xmm0 = _mm_set1_epi32((int)value);
     do
     {            
         _mm_store_si128(dst, xmm0);
